♻️ 레이아웃 구조개선
