+++ src/pages/QuestionPage.jsx
@@ -1,18 +1,56 @@
-import { useOutletContext } from "react-router";
+import { useCallback, useState } from "react";
+import { useLocation, useParams } from "react-router";
 import QuestionBox from "../components/QuestionBox";
 import FeedCard from "../components/FeedCard";
-
 import QuestionButton from "../components/QuestionButton";
+import useSubject from "../hooks/useSubject";
+import useInitialQuestion from "../hooks/useInitialQuestion";
+import useInfiniteScroll from "../hooks/useInfiniteScroll";
+import { getQuestionList } from "../api/questions";
+import QuestionLayout from "../components/QuestionLayout";
 
 const QuestionPage = () => {
-  const { questionList, questionCount, subject } = useOutletContext();
+  const [questionList, setQuestionList] = useState([]);
+  const { id } = useParams();
+  const location = useLocation();
+  const [offset, setOffset] = useState(0);
+  const [isMoreQuestion, setIsMoreQuestion] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
+
+  const { name, imageSource, questionCount } = useSubject({
+    id,
+    subject: location.state,
+  });
+  const subject = { name, imageSource };
+
+  const { isInitialLoading } = useInitialQuestion({
+    id,
+    questionCount,
+    setQuestionList,
+    setOffset,
+    setIsMoreQuestion,
+  });
+
+  const getMoreData = useCallback(async () => {
+    if (isInitialLoading || isLoading) return;
+    setIsLoading(true);
+    const { results } = await getQuestionList({ subjectId: id, offset });
+    setQuestionList((prev) => [...prev, ...results]);
+    setOffset((prev) => prev + results.length);
+    if (offset + results.length >= questionCount) {
+      setIsMoreQuestion(false);
+    }
+    setIsLoading(false);
+  }, [id, offset, questionCount, isLoading, isInitialLoading]);
+
+  const { ref } = useInfiniteScroll({ callback: getMoreData, isMoreQuestion });
   return (
-    <>
+    <QuestionLayout subject={subject}>
       <QuestionBox count={questionCount}>
         <ul className="tablet:gap-20 mt-16 flex w-full flex-col gap-16">
           {questionList.map((question) => {
             return (
-              <li key={question.id} className="">
+              <li key={question.id}>
                 <FeedCard
                   isAnswerPage={false}
                   subject={subject}
@@ -22,9 +60,10 @@ const QuestionPage = () => {
             );
           })}
         </ul>
+        <div ref={ref}></div>
       </QuestionBox>
       <QuestionButton />
-    </>
+    </QuestionLayout>
   );
 };
 export default QuestionPage;
